tests: some more hacks to improve build
